Introduced function ( getUnrarFlags() ) to determine the flags for an unrar binary. The reason for doing this is that the previous method of using an associative array is only supported in more recent versions of bash.
Because we call getUnrarFlags() from a sub shell using "exit" is ineffective so we have to
kill the parent shell using the kill program. We use SIGKILL (9) to definitely kill the shell
which isn't ideal as we can't do a clean up but this shouldn't really happen anyway because
the "kill" code should only be triggered if the programmer has made an error.