Introduced "clean up hooks" which are a family of functions (with prefix unrarall-clean- ) which can perform different clean up tasks.
The basic code to parse the command line options, verify they are valid and actually execute them (untested) are in place.

unrarall-clean-rar and unrarall-clean-nfo have been implemented but not tested.
All other hooks (except the virtual hooks "none" and "all") have stub code only and need to be implemented and tested.