Added reading passwords from file
You can now create a password list (default is ~/.unrar_passwords), which is a simple text file containing archive passwords (separated by newlines). If an archive needs a password, then unrarall will read the file in a top down manner and try to use each password until it finds the correct one or reaches EOF.

Please note, that this commit removes the ability to get asked for a password and input it manually. IMHO this is not needed, as this tool is clearly designed for mass extraction and should not wait for user input (this is also why the '-y' option is used in the code). If you need to extract an archive with a password, just put it in the passwords file  and remove the file afterwards (or just use plain unrar/7z).

This fixes arfoll/unrarall#13