Implement testing using lit and OutputCheck (both available from PyPi).
Many of the tests require the proprietary `rar` binary to run. Lit
detects if it is present at runtime and will not run tests annotated
with `REQUIRES: rar_binary`.

I've tried to write tests for many of the important features. Doing so
has revealed several bugs which we need to fix.

Unfortunately we need to rework the password support to have a command
line option so that it can be tested from lit. We do not want to be
touching the user's `.unrar_passwords` file.

In the future we could create a binary testing repo which can be dropped
in so the `rar_binary` requirement could be removed.

These tests should create a solid base for implementing continuous
integration so we can be more confident in accepting PRs and doing
our own refactorings.