Rework lit testing so that the backend to use can be specified via `--param unrarall_backend=<backend>` where valid values are `unrar` or `7z`.
This tells lit which backend is being tested which allows the backend
being tested/not being tested to be added as features. This is useful
for selectively enabling/disabling tests that only work with certain
backends.

To make testing easier the TravisCI steps to run the tests have
been moved in `tests/run.sh`.